[GTK] Implement WebAutomationSession::platformGetBase64EncodedPNGData on GTK4 and Skia https://bugs.webkit.org/show_bug.cgi?id=271985
Reviewed by Adrian Perez de Castro.

For the variant of platformGetBase64EncodedPNGData that takes a ShareableBitmap
this adds a Skia implementation.

For the variant that takes a ViewSnapshot a GTK4 implementation was added.

* Source/WebKit/PlatformWPE.cmake:
* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/UIProcess/Automation/WebAutomationSession.cpp:
* Source/WebKit/UIProcess/Automation/cairo/WebAutomationSessionCairo.cpp:
(WebKit::WebAutomationSession::platformGetBase64EncodedPNGData):
* Source/WebKit/UIProcess/Automation/gtk/WebAutomationSessionGtk.cpp:
(WebKit::base64EncodedPNGData):
(WebKit::WebAutomationSession::platformGetBase64EncodedPNGData):
* Source/WebKit/UIProcess/Automation/skia/WebAutomationSessionSkia.cpp: Copied from Source/WebKit/UIProcess/Automation/cairo/WebAutomationSessionCairo.cpp.
(WebKit::base64EncodedPNGData):
(WebKit::WebAutomationSession::platformGetBase64EncodedPNGData):

Canonical link: https://commits.webkit.org/276982@main
